Refactor store out of index.js into store/index.js
